Ensure service-worker serves root URL
create-react-app generates a service worker. This service-worker
serves up all of the static files that CRA builds, including index.html
However, it does not serve up the root, "/".

I was setting the start_url as "/", which is what I want to do, because
it looks the nicest when a user is trying to add the app to the home
screen. However, the service-worker doesn't recognize this, so the
app fails the Lighthouse audit.

There are two ideas I could think of: one is changing the service-worker,
and the other is changing the start_url. I didn't like the idea of changing
the service-worker, because I think that would involve ejecting. So
I changed the start_url and added a redirect.

I think of this as a temporary solution. I'm just playing around
with service workers for now.